# AstraServerGo.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AstraServerGo",
    "description": "基于 `AstraScheduleServerGo` 源码生成的 OpenAPI 3.0 文档。\n\n- 共 55 个 API 端点，覆盖客户端课表、天气、WebSocket、管理后台认证、用户管理、学校/年级/班级管理、配置管理、自动任务、倒数日、备份导入导出等功能。\n- JWT 认证接口使用 `jwtAuth` 安全方案（通过 POST /web/auth/login 获取 Token）。\n- 写入接口同时支持 BasicAuth（兼容旧版客户端）和 JWT + 密码验证。\n- 部分接口返回结构是动态 JSON（`map[string]interface{}`），已尽量按实际行为建模。\n",
    "version": "1.1.0"
  },
...
  "paths": {
...
    "/web/autorun": {
      "get": {
        "summary": "获取自动任务列表",
        "description": "返回所有自动任务记录，含过期/生效中/待生效状态判定。",
        "operationId": "getAutorunStatus",
        "tags": ["web-autorun"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "自动任务列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/AutorunRecordView" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": []
      }
    },
    "/web/autorun/hash/{hashid}": {
      "get": {
        "summary": "按 HashID 查询自动任务",
        "description": "无记录时 data 为空数组。",
        "operationId": "getAutorunHashStatus",
        "tags": ["web-autorun"],
        "parameters": [
          {
            "name": "hashid", "in": "path", "description": "自动任务 HashID", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "查询结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "oneOf": [
                        { "type": "array", "items": { "$ref": "#/components/schemas/AutorunRecordView" } },
                        { "$ref": "#/components/schemas/AutorunRecordView" }
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": []
      }
    },
    "/web/autorun/{hashid}": {
      "delete": {
        "summary": "删除自动任务",
        "description": "需要 BasicAuth 认证。",
        "operationId": "deleteAutorunRecord",
        "tags": ["web-autorun"],
        "parameters": [
          {
            "name": "hashid", "in": "path", "description": "自动任务 HashID", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "deleted", "id"],
                  "properties": {
                    "status": { "type": "integer", "example": 200 },
                    "deleted": { "type": "integer", "format": "int64" },
                    "id": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Basic Auth 认证失败" },
          "404": {
            "description": "记录不存在",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/autorun/compensation": {
      "put": {
        "summary": "新增/更新调休规则",
        "description": "type 字段会被服务端强制修正为 0。需要 BasicAuth 认证。",
        "operationId": "putCompensationRule",
        "tags": ["web-autorun"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CompensationRulePayload" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "自动任务写入成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AutorunPutSuccess" }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          },
          "401": { "description": "Basic Auth 认证失败" },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/autorun/timetable": {
      "put": {
        "summary": "新增/更新换作息规则",
        "description": "type 字段会被服务端强制修正为 1。需要 BasicAuth 认证。",
        "operationId": "putTimetableRule",
        "tags": ["web-autorun"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TimetableRulePayload" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "自动任务写入成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AutorunPutSuccess" }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          },
          "401": { "description": "Basic Auth 认证失败" },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/autorun/schedule": {
      "put": {
        "summary": "新增/更新换课规则",
        "description": "type 字段会被服务端强制修正为 2。需要 BasicAuth 认证。",
        "operationId": "putScheduleRule",
        "tags": ["web-autorun"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ScheduleRulePayload" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "自动任务写入成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AutorunPutSuccess" }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          },
          "401": { "description": "Basic Auth 认证失败" },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/autorun/all": {
      "put": {
        "summary": "新增/更新组合规则（作息+课表）",
        "description": "type 字段会被服务端强制修正为 3。需要 BasicAuth 认证。",
        "operationId": "putAllRule",
        "tags": ["web-autorun"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AllRulePayload" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "自动任务写入成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AutorunPutSuccess" }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          },
          "401": { "description": "Basic Auth 认证失败" },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ScopeField": {
        "description": "作用域字段，支持单个 scope 字符串（school/grade/class）或字符串数组，也可为空表示 ALL。",
        "oneOf": [
          { "type": "string" },
          { "type": "array", "items": { "type": "string" } }
        ]
      },
      "AutorunRecordView": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "HashID" },
          "type": {
            "type": "string",
            "description": "任务类型：COMPENSATION(0) | TIMETABLE(1) | SCHEDULE(2) | ALL(3)"
          },
          "priority": { "type": "integer" },
          "status": {
            "type": "string",
            "description": "任务状态：待生效 | 生效中 | 已过期"
          },
          "scope": {
            "type": "array",
            "items": { "type": "string" },
            "description": "作用域列表"
          },
          "content": {
            "type": "object",
            "additionalProperties": true,
            "description": "规则内容（根据 type 不同结构不同）"
          }
        }
      },
      "SchedulePeriod": {
        "type": "object",
        "required": ["no", "subject"],
        "properties": {
          "no": { "type": "integer", "minimum": 1, "description": "课节序号" },
          "subject": { "type": "string", "description": "科目简称" }
        }
      },
      "CompensationRulePayload": {
        "type": "object",
        "properties": {
          "type": {
            "type": "integer",
            "description": "输入会被服务端修正为 0"
          },
          "scope": { "$ref": "#/components/schemas/ScopeField" },
          "priority": { "type": "integer" },
          "id": { "type": "string", "description": "可选；不传时生成" },
          "content": {
            "type": "object",
            "required": ["date", "useDate"],
            "properties": {
              "date": { "type": "string", "format": "date", "description": "节假日日期" },
              "useDate": { "type": "string", "format": "date", "description": "对应的补班工作日" }
            }
          }
        }
      },
      "TimetableRulePayload": {
        "type": "object",
        "properties": {
          "type": {
            "type": "integer",
            "description": "输入会被服务端修正为 1"
          },
          "scope": { "$ref": "#/components/schemas/ScopeField" },
          "priority": { "type": "integer" },
          "id": { "type": "string" },
          "content": {
            "type": "object",
            "required": ["date", "timetableId"],
            "properties": {
              "date": { "type": "string", "format": "date" },
              "timetableId": { "type": "string", "description": "作息表名称" }
            }
          }
        }
      },
      "ScheduleRulePayload": {
        "type": "object",
        "properties": {
          "type": {
            "type": "integer",
            "description": "输入会被服务端修正为 2"
          },
          "scope": { "$ref": "#/components/schemas/ScopeField" },
          "priority": { "type": "integer" },
          "id": { "type": "string" },
          "content": {
            "type": "object",
            "required": ["date", "schedule"],
            "properties": {
              "date": { "type": "string", "format": "date" },
              "schedule": {
                "type": "object",
                "required": ["periods"],
                "properties": {
                  "periods": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/SchedulePeriod" }
                  }
                }
              }
            }
          }
        }
      },
      "AllRulePayload": {
        "type": "object",
        "properties": {
          "type": {
            "type": "integer",
            "description": "输入会被服务端修正为 3"
          },
          "scope": { "$ref": "#/components/schemas/ScopeField" },
          "priority": { "type": "integer" },
          "id": { "type": "string" },
          "content": {
            "type": "object",
            "required": ["date", "timetableId", "schedule"],
            "properties": {
              "date": { "type": "string", "format": "date" },
              "timetableId": { "type": "string" },
              "schedule": {
                "type": "object",
                "required": ["periods"],
                "properties": {
                  "periods": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/SchedulePeriod" }
                  }
                }
              }
            }
          }
        }
      },
...
      "ErrorWithErrorField": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "DetailResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string" }
        }
      },
...
      "AutorunPutSuccess": {
        "type": "object",
        "required": ["status", "id"],
        "properties": {
          "status": { "type": "integer", "example": 200 },
          "id": { "type": "string" }
        }
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "（已废弃，仅供旧版客户端兼容）用户名：`AstraSchedule` 或 `ElectronClassSchedule`，密码：配置文件 `secret.token`"
      },
      "jwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token，通过 POST /web/auth/login 获取。写入接口还需在 X-Verify-Password 头中提供密码。"
      },
      "internalAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Internal-Secret",
        "description": "内部 API 密钥，用于服务间调用。配置文件 [internal].secret"
      },
      "regToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Reg-Token",
        "description": "注册 JWT 令牌，由 reg-to 签发，有效期 10 分钟。包含完整注册信息。"
      }
    }
  },
  "servers": [
    {
      "url": "https://{hostvar1}.getastra.cn"
    }
  ],
  "security": []
}